+++ lib/capistrano/recipes/deploy/remote_dependency.rb
@@ -1,15 +1,14 @@
 require 'capistrano/errors'
+require 'capistrano/recipes/deploy/remote_dependency'
 
 module Capistrano
   module Deploy
-    class RemoteDependency
-      attr_reader :configuration
+    class RemoteDependency < Dependency
       attr_reader :hosts
 
       def initialize(configuration)
-        @configuration = configuration
-        @success = true
         @hosts = nil
+        super(configuration)
       end
 
       def directory(path, options={})
@@ -33,25 +32,6 @@ module Capistrano
       def command(command, options={})
         @message ||= "`#{command}' could not be found in the path"
         try("which #{command}", options)
-        self
-      end
-
-      def gem(name, version, options={})
-        @message ||= "gem `#{name}' #{version} could not be found"
-        gem_cmd = configuration.fetch(:gem_command, "gem")
-        try("#{gem_cmd} specification --version '#{version}' #{name} 2>&1 | awk 'BEGIN { s = 0 } /^name:/ { s = 1; exit }; END { if(s == 0) exit 1 }'", options)
-        self
-      end
-
-      def deb(name, version, options={})
-        @message ||= "package `#{name}' #{version} could not be found"
-        try("dpkg -s #{name} | grep '^Version: #{version}'", options)
-        self
-      end
-
-      def rpm(name, version, options={})
-        @message ||= "package `#{name}' #{version} could not be found"
-        try("rpm -q #{name} | grep '#{version}'", options)
         self
       end
 
@@ -85,15 +65,6 @@ module Capistrano
         self 
       end
 
-      def or(message)
-        @message = message
-        self
-      end
-
-      def pass?
-        @success
-      end
-
       def message
         s = @message.dup
         s << " (#{@hosts})" if @hosts
@@ -102,8 +73,7 @@ module Capistrano
 
     private
 
-      def try(command, options)
-        return unless @success # short-circuit evaluation
+      def try!(command, options)
         configuration.invoke_command(command, options) do |ch,stream,out|
           warn "#{ch[:server]}: #{out}" if stream == :err
           yield ch, stream, out if block_given?
